refactor TaxOptions to their own utils
